updating reader
